# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse:
    properties:
      error:
        example: unauthorized
        type: string
      request_id:
        example: abc123
        type: string
    type: object
...
  github_com_Yab1_golang-template_internal_platform_httpx.Meta:
    properties:
      pagination:
        $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.Pagination'
      version:
        example: "1"
        type: string
    type: object
  github_com_Yab1_golang-template_internal_platform_httpx.ObjectResponse:
    properties:
      meta:
        $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.Meta'
      result: {}
      status:
        enum:
        - success
        example: success
        type: string
    type: object
  github_com_Yab1_golang-template_internal_platform_httpx.Pagination:
    properties:
      limit:
        example: 20
        type: integer
      next_cursor:
        example: eyJ0IjoiMjAyNi0wOS0xOFQxMjowMDowMFoiLCJpIjoiLi4uIn0
        type: string
      offset:
        example: 0
        type: integer
      total:
        example: 42
        type: integer
    type: object
  modules_file.UploadResponse:
    properties:
      content_type:
        type: string
      driver:
        type: string
      key:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: yeabsera.dev@gmail.com
    name: Yeabsera
    url: https://yeabsera-dev.vercel.app/
  description: 'REST API template: JWT access+refresh, RBAC, posts CRUD, files (local/minio/s3),
    mail, Redis rate limiting, human reference ids.'
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.google.com/policies/terms/
  title: Golang Template API
  version: 0.0.1
paths:
...
  /files:
    post:
      consumes:
      - multipart/form-data
      description: Multipart form field "file". Backend is local, s3, or minio via
        STORAGE_DRIVER.
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ObjectResponse'
            - properties:
                result:
                  $ref: '#/definitions/modules_file.UploadResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a file
      tags:
      - files
  /files/{key}:
    delete:
      parameters:
      - description: Object key
        in: path
        name: key
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a file
      tags:
      - files
    get:
      description: Streams from local disk, or redirects to a presigned S3/MinIO URL
      parameters:
      - description: Object key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
      summary: Get a file
      tags:
      - files
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access JWT (or AUTH_DEV_TOKEN
      in development).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"